+++ src/pages/userPage/helpers.js
@@ -1,18 +1,13 @@
-// code from formik lib
+export const getInputClassName = (error, touched) => {
+  if (!error && !touched) {
+    return ''
+  }
 
-import React from 'react'
+  if (!error && touched) {
+    return 'success'
+  }
 
-export const DisplayFormikState = (props) => (
-  <div style={{ margin: '1rem 0' }}>
-    <h3 style={{ fontFamily: 'monospace' }} />
-    <pre
-      style={{
-        background: '#f6f8fa',
-        fontSize: '.65rem',
-        padding: '.5rem',
-      }}
-    >
-      <strong>props</strong> = {JSON.stringify(props, null, 2)}
-    </pre>
-  </div>
-)
+  if (error && touched) {
+    return 'error'
+  }
+}
